requirements.txt
asgiref==3.8.1
Django==4.2.11
django-crispy-forms==2.1
gunicorn==22.0.0
packaging==24.0
pillow==10.3.0
sqlparse==0.5.0
typing_extensions==4.11.0
whitenoise==6.6.0

requests~=2.32.3
reportlab~=4.2.0